refactor(auth): extract middleware seam
Move AuthMiddleware, _dispatch_builtin, _dispatch_proxy, plus
the middleware-owned constants _PUBLIC_PATHS and _LOGOUT_PATH
to src/houndarr/auth/middleware.py.  auth/__init__.py collapses
to a pure re-export shim: ~130 lines of imports + __all__ +
__getattr__ pass-through for the three state-bearing globals.

The middleware calls every seam helper via its module (e.g.
``proxy_auth._is_trusted_proxy(request)`` rather than a direct
``from .proxy_auth import _is_trusted_proxy`` binding) so tests
that monkeypatch the module attribute propagate the patch into
the middleware without touching the import site.  The Phase 1
ordering pin moves its patch target to
``houndarr.auth.proxy_auth`` to match.
